KEYMOD_* undefined
For some strange reason the KeyEvent.DOM_VK_* constants are undefined
in gemseBase.js where the KEYMOD_* constants get set. Actually, in
inputtest.xhtml, this does not happen.
However, I replaced the KeyEvent.DOM_VK_* with their actual value.

Also, the KEYMOD_* constants must be strings, not numbers, so the
conversion to a string is necessairy. (This is correct in
inputtest.xhtml.)