Prevent closing if changes are not saved
The history object now also tracks wich state is saved to disk. On
loading and saving the history object has to be informed. Clsoe uses
the history object to find out whether it can close. It has a
parameter force. If it is set to true, it does always close.