Fixed undo/redo bug
A bug in applyTreePointer() of the Change class caused the array
object pointer (passed as an argument) to be changed. The caller
expects it to stay the same. (I mistakenly thought that arrays are
handed to a function by value, but since they are normal objects, they
are handed over by reference of course. So I must not use shift().)

I also added code in recordBefore of the Change class that tries to
detect bugs in deriveTreePointer and applyTreePointer.