Solved access rights problem at start
When the editor is loaded outside the chrome, it has no access rights
to the components. This means that URI handling fails. Since the
editor tried to find the current working directoy at start, it fails
loading. Now the workingDirectory property is not set at start, but is
now a getter which finds the working directory on request.