Worked around privilege restriction when loading
When the user loads an equation, the a relative URI is transformed to
an absolute one using the workingdirectory as base if the privileges
for accessing XPCom components. In such a case, XMLHttpRequest does
transform the URI to an absolute one, perhaps using the location of
the editor.xhtml file as base.