Added overlay for browser: edit formulas via popup
An overlay for the browser has been added. For this, a gemse.manifest
file is now included in the repository. A user can now rightclick on
a math element and choose "edit in Gemse". Actually, if he clicks a
descendant of a MathML element, the math element gets opened in a
new Gemse editor. If he clicks an element that is an ancestor of one
or several math elements, all of them get opened in a new Gemse
editor. If the user has made a selection, all selected or partially
selected nodes are opened.
Have fun with this powerful feature!