Changed handling of movement commands
The movement commands of the edit mode have been adapted such that
they do not change the cursor themselves. Thes recieve an element as
argument and return a destination element. Like that, the visual mode
is also able to use edit mode movement commands. However, it does that
only if the movement command moves along siblings.