Separated command tables and constants from rest
This is in order to ease configuration for the user. Unfortunately, if
one wants to load the command tables, one has to previously load the
mode itself. This is because the functions that are referenced in the
tables have to exist. This is a drawback for the automatic
documentation generation system.