Automativ documentation: command placeholders
In the command description one sometimes wants to use the associated
command inside a kbd element. This is especially useful when giving
examples. However, since the user can change the commands and since
the documentation has to adapt automatically, one can use commands in
the description only with placeholders. The element for that is cmdph
in the internal namespace. If it has no attributes and turns up in the
description of a command, this command will be inserted.
Alternatively, it can have a href attribute, which has the function
name (i.e. the id) of a commadn as value. In this case the referenced
command will be inserted.