Edit mode receives new cursor from called mode
If a mode removes the element to which is the cursor of the edit mode,
it can now give the edit mode a new cursor when it terminates. This
fixes a bug which made problems when selecting a placeholder (which
was left over from a last call to insert mode) and inserting before
it.