clearing of input buffer in finally
Up to now it happened from time to time that you can no longer clear
the input buffer, because its current content causes an error to be
thrown. Now the code that tests for an escape character at the end of
the input buffer have been moved into a finally block, so it is
executed even if an error is thrown.