insertIn now reports better/correct changed node
Mistakenly, insertIn always reported the parent node of the cursour as
the element that gets changed. This is not only inefficient, it also
causes an error when inserting into the root element, i.e. the math
element. This is especially nasty if you start with an empty math
element as formula.