Extended input substitution
Now the input substitution table is built using
http://www.w3.org/2003/entities/2007/w3centities-f.ent
which is part of
http://www.w3.org/2003/entities/2007doc/
Unfortunately, these entity names are not prefix free, so one can not
access all characters using \foo. So I added a syntax similar to XML
entities, wich always explicitely marks the end of a string that has
to be replaced.

The current implementation has some drastic problems: It does
substitution as long as it finds signs that trigger a substitution.
So, if one of those substitutions have such a sign in the replacement
string, this sign will also trigger a substitution. This is not what
the user expects.