Input substitution simplified
For the sake of simplicity of the code, the input substitution has
been rewritten such that it works much like entities in XML. It is
possible to insert characters by codepoint using decimal or
hexadecimal notation (for example &#8658; or &#x21d2; respectively).
Also, one can use one of the entries in the inputSubstitutionTable
(for example &DoubleRightArrow;).