Fixed bug reintroduced in last merge
In the last merge, a bug has been reintroduced: It was not possible to
insert into the math element, because the change object was given the
parent node of the math element as changed element. However, the
faulty code was somewhere else this time. The member function of the
edit mode that is used to start an insert mode used this.cursor at
some places instead of the argument cursorInElement.