Fixed nasty memory problem
The input substitution table was loaded by code not enclosed in some
function. This caused the variables (which are declared by var
statements) to stay in global scope and were never destroyed. This
caused firefox to use around 800MB of memory. This is fixed by putting
the code in a function.
Howver, it is still a riddle to me why it needs so much memory. It may
even be that this fix does solve the problem only partially, since it
could be, that now Gemse uses still so much memory at startup.