Huge memory problem in inputSubstitution solved
When loading the input substitution table, Gemse used a huge amount of
memory, about 1GB! After loading the memory was freed. It turned out
that one must not use request.responseText multiple times. One should
make first a full copy of the response text and get rid of the
response object. I am not shure why exactly this memory problem
occures. It may be that the repeated application of a Regex on
request.responseText is the problem.
I also checked ucd.js. It does not have this problem. But I added a
delete for the request object in the hope to use a little bit less
memory.