UString integrated in String
It has turned out that using separate objects for UString and String
is bad, since while looking at the code, one is never sure of which
kind an object actually is. Even more so, because the methods of
UString mostly return string values.

Now, a much better solution has been implemented. The methods of
that where before methods of UString.prototype have been added to the
prototype of String directly, using another name (which begins with u)
in order to not clash with the UTF16 based variants. So, both, the
UTF16 based functionality and the Unicode character based one are
always available on every string.