Use features from UString in editMode
The editMode has been modified such that unicode characters from
higher planes in the input are handles in a correct way. This was easy
todo, since only the input handler had to be adapted. Now it is
possible to have characters outside the BMP as commands or single
character parameters.
Of course, everything still has to be tested.