Added exporter functionality
The command :export can be used to transform into another format, like
TeX. The TeX exporter is the only one implemented so far. It uses the
XSLT MAthML Library.
(see http://www.raleigh.ru/MathML/mmltex/index.php?lang=en)
It must be placed in the folder approach1/exporters/xsltml to work.
At the moment, the folder exporters contains a symbolik link to to
the untracked folder xsltml called xsltml, so it works for me.
However, xsltml will not be shipped with Gemse for now, because of
lizence issues.