UCD: Handle unicode combining characters
ucd.js now provides some information about combining characters:
Their position and their standalone variant. The position is derived
from the combining class, which is rather heuristic since combining
classes are not normatively bound to a position (as far as I know).
A character that has 0020 plus a character c as decomposition, it is
considered to be the standalone version of c.

Since these information from the UCD is perhaps not enough for
mathematical purposes, such information can be added manually to the
file UCD/Gemse_Combining.txt without the need of modifying
UCD/UnicodeData.txt directly.

The unicode insert mode (ucdInsertMode) uses this information to place
an accent on the preceding element when the user inserts a combining
character. However, this feature is not yet fully implemented, only
mover elements are used.