Documentation: Command list processor
Implementations not in the global scope (that is, not part of the
window object) have not been recognized. This has been fixed by
getting the implementations when looking at the bindings if they have
not been found before. Of course this requires that a binding exists
for the command in question.

In the documentation, one can use the element cmdph to refere to a
command which is transformed into a string by the command list
processor. Since a command can have more than one binding, one
sometimes wants to refer to a binding with some given properties. This
is now possible with the select attribute. Its value is a JavaScript
expression resulting in a boolean. It must use the variable "b" which
is a binding object.