Register becomes RegisterData and RegisterManager
The Register Object is now named RegisterData and holds data received
from a register or to be stored in a register. The actual handling of
registers is now done by the RegisterManager class. The editor object
holds an instance of RegisterManager in its registerManager property.
Registers can no longer be set by accessing editor.registers,
the get/set methods of editor.registerManager has to be used.

This breaks commands that use registers!