New directory structure
Moved around a lot of files into a more useful directory structure.
The reason for this is that ther are some differences between the
XULRunner application and the Firefox extension.

New structure:
src_common: Files common to both, the XULRunner application and the
Firefox extension. It mainly contains chrome/content.
src_XULRunnerApplication: Special files for the XULRunnerApplication
src_FirefoxExtension: Special files for the Firefox extension
test_FirefoxExtension: Collects all files for firefox using symbolic
links. This directory is referenced from the firefox's extensions
folder of my profile.
test_XULRunnerApplication: Collects all files for the XULRunner application
using symbolic links to ../src_*. (application.ini is an exception, it
must be a copy of ../src_XULRunnerApplication/application.ini) This
directory is used to test the XULRunnerApplication.