Removed link to the exporter xsltml
since packaging does resolve symbolic links and xsltml should not be
included in the Gemse packages because of license issues. (It has not
yet been decided under which license Gemse will be available.)
Does the link have to be added because of testing purposes, this
should be done in the text_* directories which never get packaged into
a zip file. (They get packaged into a tar file, but tar files do not
resolve symbolic links.)

Additonally. a readme file has been added to the exporters folder in
order to preserve it in git and in order to help users installing
xsltml in Gemse manually.