Views: viewport, viewClass, viewset
This is a big change. The document no longer contains a "pool" which
contains a "container" which in turn contains the equation and an
element for every view.
Equations no longer turn up in the document directly (although at most
one view may do that anyway). A set of views are put in the document
according to a predefined viewset that can be chosen by the user.
These views show the equation in some way.

Some problems must be fixed. editor.xhtml has not yet been adapted.