Fixed bug in views causing weird behaviour
Views which show stuff of the context node as indicated by the mode
went on, even if the mode returned null. Now they stop the building
process if that is the case. This bug caused some weird behauviour: In
certain situations in the input node, a placeholder was not replaced,
the cursor was just moved behind it.

Additionally I added a comment that I still have to implement that
hidden view are not built.