Added MessageView and fixed others
equationEnv.notificationDisplay has been replaced by
editor.showMessage. This method takes a string or a DOM element. It
puts what it gets into editor.lastMessage as a DOM element. The
MessageView puts this element into the document.

The DirectView has been fixed: Bevore it always put the equation into
the document again, now it should do it only if the root element has
been replaced, which is as intended from the beginning.