mrowEnvelop creates apply depending on first child
The mrowEnvelop command of the edit mode (usually bound to the space
bar) now creates an apply element instead of an mrow element if the
first element of the selection is a content MathML element.

However, at the moment it is not possible to really test whether an
element is a Content MathML element. So the implemented algorithm is a
hack: If the name of an element begins with "m" or the element is
found in the operator descriptions, it is assumed to be a Presentation
MathML element.