Added command followRef
If the element under the cursor has an xref attribute, the cursor is
moved to the element having an id or xml:id attribute with the same
value.
If there is no xref attribute, but an id or xml:id attribute, the
cursor is moved to an element with an xref attribute with the same
value. In other words, it follows the reference backwards.