Other strategy for updating views
The views are now updated (this is, editor.viewsetManager.build() is
called) after the input handler loop has finished and something has
been executed. This means that the
modes no longer update the views themselves. This prevents
unnecessairy multiple updates.

The method equationEnv.updateViews has been removed and must no longer
be used.

Adapted all modes to new view update strategy

That is, removed all calls to equationEnv.updateViews and removed
unnecessairy calls of mode.moveCursor(mode.cursor) which were just
used to update the views.

Some calls of mode.moveCursor in the init methods of modes were there
so that the cursor is shown right after calling the mode. They have
been replaced by mode.showCursor().

The edit mode did not hide its cursor when calling a mode and it did
not show the cursor when the mode returned. This is now fixed as well.