Topmost vbox now has max-height and scrollbar
The topmost vbox (that it, the viewsetDock) has now max-heigth:100%.
This makes sure that the input box is always visible. Additionally,
the topmost vbox has now overflow:auto which allows the user to scroll
down via mouse if a view is too high to fit into the window.

This change is not really a solution to the problem. It should never
happen that you need to scroll with the mouse. However, I don't have a
better solution at the moment.