Fixed input substitution
Before, the DTD containing the entities was parsed directly at startup
of Gemse. Furthermore, it was parsed the wrong way: The entities have
to be decoded twice, once is not enough.
Now, a perl script has been added which can be used to transform the
DTD into a tabulator/newline separated list where the values of
entitites are already correctly decoded. This list is then parsed by
Gemse at startup.