Error handling in input handler, showMessage
The input handler of the editor now catches thrown errors and shows
them to the user by calling editor.showMessage.

editor.showMessage has been improved. It shows objects of class Error
in a useful way. DOM nodes are still shown as is. All other objects
are converted to a String using their toString method and the result
is presented to the user. Styles have been added that style Errors in
red.

A global option detailedErrors has been added. If it is set to true,
showMessage add a backtrace along every error message, which includes
file name and number of the line where the error occured.