Added killPrevious command to all insert modes
The command killPrevious removes the element right before
the cursor. If there is no element before the cursor, the
element in which the cursor resides is removed. If there are some
surrounded elements, all of them are removed.

I have thought about the following improvement: Do only remove the
leaf before the cursor, not the whole element. So, if the cursor is
behind an mrow, remove the last element of the mrow and place the
cursor in the mrow. This perhaps would be natural, after the user used
the coursor jump command (Enter) to jump behind the element the cursor
is currently in. This has not yet been implemented. It will be
implemented if tests show that it could be useful.

The command table from the UCD insert mode references the
implementation from the trivial insert mode. (Only works because both
modes have the same interfaces for handling cursors.) The content
insert mode contains a copy of the implementation from the trivial
insert mode, since it must stay independent.