Change element when calling insert mode
When the edit mode calls an insert mode, it has to record a change so
that undo and redo work. For this, it has to indicate the root element
of the subtree of the formula that is going to be changed. Up to now,
this has been the element the cursor is in. Since edit modes can do
modifications outside this subtree (by the killBefore command or by
cursor jumping or whatever), this causes undo/redo to mess up
Gemse completely.

Now the math element is used as root. Drawback: Undo/redo uses lots of
memory for small changes made by an insert mode.