Added java class URLSetPolicy
It is needed to run jomdoc and its other libraries with elevated
privileges. This is not because it needs to access files in its own
jar files since the URLClassLoader allows that anyway. It seems that
it is needed by Saxon to register itself globally.

The class file is included in the repository as well, so one does not
need to recompile when switching branches.