Added workaround for JOMDocs class loader bug
At the time of writing, the class org.omdoc.jomdoc.ntn.coll.ntn.B from
JOMDoc uses the system class loader to load the bundled files
(Bug #603 of JOMDoc). This does not work, because Gemse uses a special class
loader for JOMDoc. This commit works around by imitating
the constructor of the class B. (Note that the methods the constructor
calls also use a ClassLoader to locate files, but always using the
ClassLoader of the respective class instead of the system
ClassLoader.)