Clear input buffer if errror is thrown (#13)
If the loop over the input is terminated because an error is thrown,
valid or invalid commands remain in the input buffer. However, when
the loop terminates, the inputBuffer must not contain any further
complete commands.

This issue is now (temporarly) by clearing the input buffer when an
error is thrown.