NTNView no longer renders after cursor movement
The NTNView no longer renders the formula using JOMDoc if only
internal attributes have changed but nothing else. This mainly means
that no rendering happens after moving the cursor. In order to
correctly show the cursor, the old result of the renderer is compared
to the new one and internal attributes are updated accordingly.

This feaature can be turned of by setting the option NTNView.shortcut
to false.