Skipping input events if there are more
Input events are now skipped if the user has already generated more
of them. This is achieved using a timer. If there are no more input
events in a very short period of time, a command event is issued using
a timer. On the other hand, every input event resets the timer.

The timer is set to 10ms (see ui.xml). Perhaps this value has to be
adapted later on. The input box is implemented in a binding in ui.xml
wihch is attached using a rule in gemse.css.