Equation now kept in its own document
EquationEnv now keeps the equation inside its own document.
This document object is stored in the property document of the
EquationEnv instance.

The problem that turns up now is that always this document has to be
used when creating elements for the equation or when searching inside
the equation. (Although it seems to work anyway sometimes.)

The mml_ functions now no longer require the root element to be a math
element.