Fixed owner document issues
Since the equation is now kept in its own document, a lot of code that
uses window.document to create new nodes or to traverse the equation
breaks. These problems should all have been fixed in this commit by
using the document of the equation instead.
In order to facilitate this, all modes have a new property d which
is set to the document of the current equation.

Be aware that many occurences of window.document are correct, since the edit
mode often has to generate a message that is shown in a view.
In one place it is even very important that window.document is used.
There, document has been replaced with window.document to emphasize
the situation.