Support for Mozilla 2
Increased maximal version of Firefox and XULRunner in install.rdf and
application.ini.

Manifests were updated to acommodate new registration process for
XPCOM components. Rewrote XPCOM component for command line handling.
(Uses now XPCOMUtils.jsm)

The location for the manifest has changed in XULRunner 2. It now
resides in the same location as for Firefox. For backwards
compatibility the new manifest for XULRunner just links to the old
one.

Detection of the installation directory had to be updated to use the
new AddonManager instead of the extension manager on Mozilla 2.