The command newEquation can now create any element
The command newEquation now creates an OpenMath OMOBJ element when
"om" as argument is given or a MathML math element when "m" is given.
If an argument containing a whitespace is given, the part before the
whitespace is considered the namespace and the part after it the local
name of the root element of the new equation. The namespace can also
be the name of a prefix recognized by the default namespace resolver
defined in const.js
If no argument is given, the default equation is created, that is,
null is handed to the method newEquation of the editor.

Documentation is included.

This feature will have to be changed in the future: A user most
probably wants to define her own templates for new equations.