Properly using AddonManager.getAddonById callback
Before, initialisation of GemsePEditor was completely done without
waiting for AddonManager.getAddonById calling back. Because of this,
the installation directory of the addon or XULRunner application was
unknown during initialisation. This caused problems when an equation
loaded when starting Gemse caused the Java subsystem to be loaded,
which is only possible when knowing the installation directory. It
also could cause other problems in the future.

Now the initialisation consists of two steps, the second one invoked
by the callback. The constructor of GemsePEditor now also takes a
callback as an argument which is called after initialization has been
completed. The editor object must not be used until this happened.