Views: Reading optoins from viewport at right time
Options set on the viewport element are now read by the constructors
of the Views. It is not possible to do that before the View has been
instantiated and after running the constructor it is too late, since
it might need the options. This is the case for the NTNView. Setting
Options locally for an NTNView (e.g. using the viewsetconf) works now.