Restore command input
Gemse uses a hybrid way of getting keyboard input: On the one hand, it
allows the user to enter text, which also allows for composition of
single characters with multiple key strokes: For instance, gemse can be
used together with a math input mode provided by the operating system.
On the other hand, it allows key combinations for special commands. This
is tricky to do right, but combining the default behaviour of a textarea
with a custom key handler works out. The original implementation used
Firefox-specific interfaces.

To fix this, the following had to be done:
- Use a HTML textarea instead of a XUL textbox, which also required some
  CSS changes.
- Fix the custom event handler due to changed representation of keys.