Implement loading and saving of files
This implementation uses a file picker input element for loading and
navigation to a blob URL for saving. This is both suboptimal.
Especially, saving requires the user to manually choose the correct
file. That is, there is only a "save as" and no way to do a normal
"save" of the opened file.

This implementation will have to be replaces as soon as the File API is
implemented in Firefox.