Make individual parser procedures functional
Each of these procedures parse a part of a command:
- The prepended number indicating repetition
- Pre-arguments such as `"a` to indicate a registory to copy to
- The command itself
- Any additional command arguments

Each of these procedures are now functional. That is, their behaviour is
fully defined by the arguments provided, they do not access or modify
any other state and they do not modify any values passed to them. This
change will make it much easier to combine these procedures together.