Move value properties from prototype into object
The properties moved from the prototype of CommandInstance to the
instances of CommandInstance were defaults in some sense: Their values
could be different for different instance of CommandInstance. Mostly,
however, they would remain the default value. Therefore, the default had
been set on the prototype while values deviating from the default had
been set on each individual instance of CommandInstance.

This is a bad strategy for several reasons:
- If the default is not immutable, one can mistakenly change the default
  on the prototype.
- Compilers rely on objects always having the same structure for the
  same piece of code, which implies that always the same properties must
  be set on the objects directly. Here however, each object has set just
  those properties that differ from the default. Which properties those
  are differs from object to object.
- Serialization and iteration of properties is difficult because some
  properties originate form the objects, while others from its
  prototype.

Therefore, all properties are now set in the constructor to their
default on every instance of CommandInstance. Only method calls are
delegated to the prototype, as the methods are always the same.