Rewrite parser in a functional manner
The parser is now a single exported function which only uses the passed
arguments to create a new CommandInstance.

As a result, the complete command parser is now functional. It no longer
carries state. Therefore, calling code no longer needs to create an
instance of the parser and set options on this instance. Instead,
calling code must do a normal function call and pass all required values
and options on every call.

This simplifies state management in the modes as the state of the
command parser no longer needs to be carried around and potentially
updated. A small drawback is that each mode now must take care of state
changes, such as removing an executed command from the input buffer.