Load module dynamically without module script
Chrome does not support script elements with `type="module"` in
XHTML.[1] In order to get the application running in Chrome too, we use
a normal script element. A normal script cannot import modules
declaratively, only modules can do that. But it is possible to load a
module dynamically, which we now do.

The application now finally works in Chrome too.

[1] https://issues.chromium.org/issues/40518469