Unify indentation in examples of introduction
The MathML introduction used inconsistent indentations in the example
MathML snippets. This has been fixed. At least one instance of bad
indentation may have been intentional to higlight the difference to
another snippet. However, I do not think this was a good idea, so it has
been unified as well.